+++ blockchain/block.js
@@ -1,6 +1,6 @@
 const hexToBinary = require('hex-to-binary')
-const { GENESIS_DATA, MINE_RATE } = require('./config');
-const cryptoHash = require('./crypto-hash');
+const { GENESIS_DATA, MINE_RATE } = require('../config');
+const cryptoHash = require('../util/crypto-hash');
 
 class Block {
   constructor({ timestamp, lastHash, hash, data, nonce, difficulty }) {
