+++ util/crypto-hash.js
